Add caching of static files through Cloudflare
The default Netlify cache-control header, are
"max-age=0, must-revalidate, public"

This is a problem as cloudflare cannot cache the content.

This fixes it by moving the content to a ``Cache-Control: public,
s-max-age=604800``. This takes precedence over max-age or the Expires header,
but it only applies to shared caches (e.g., proxies) and is ignored by a
private cache. This allow netlify to tell cloudfalre to cache assets for 7
days (css/images/js).